tools/gemini-cli/pyproject.toml
[project]
name = "gemini-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "google-genai>=1.61.0",
    "rich>=14.3.2",
]
